+++ test/support/plug_router.ex
@@ -11,7 +11,7 @@ defmodule MyRouter do
 
   plug(Cldr.Plug.SetLocale,
     apps: [:cldr, :gettext],
-    from: [:path, :query],
+    from: [:path, :query, :assigns],
     gettext: TestGettext.Gettext,
     cldr: TestBackend.Cldr
   )
@@ -21,6 +21,10 @@ defmodule MyRouter do
   plug(:dispatch)
 
   get "/hello/:locale" do
+    send_resp(conn, 200, "world")
+  end
+
+  get "/hello", assigns: %{cldr_locale: "fr-FR"} do
     send_resp(conn, 200, "world")
   end